+++ noodles-util/src/variant/async.rs
@@ -1,3 +1,4 @@
 //! Async variant format utilities.
 
+pub mod fs;
 pub mod io;